Purely attribute-driven ossec.conf and agent.conf
Addresses #31.

It probably goes without saying that this is a totally breaking
change. It is so far removed from the previous approach that I do not
think it is worth trying to migrate existing users as almost all of
them would have written their own configuration files anyway.

Unfortunately Gyoku cannot pretty print its output at present. This is
pending a long-standing change to Builder.