Fix Failing CI
Right now Test Kitchen is just looking for an enabled service but
there seems to be an issue with the test VMs and enabling the service
without a reboot.

This can be solved in another change, for now, just verify that the
service was installed. The test coverage will also need to be expanded
in the future.

Fixes #98